[helios][fx] Add get_edge_label to FxGraphDrawer (#80481)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80481

Allow customizing the label used for an edge in the FxGraphDrawer
with a new overridable function `_get_edge_label`. This function takes
in the source node and destination node (called `node` and `user`). The
GraphModule is provided to access submodules if necessary.

If this function returns `None` (like the default implementation), then no
label is added to the edge.

Differential Revision: D37352431

fbshipit-source-id: 7ee72af5dbf61246ef157f308b533c86d6b186d8